news_scrapper/requirements.txt
requests
crawl4ai
beautifulsoup4
lxml
feedparser
python-dateutil
